+++ ceammc/CHANGELOG.md
@@ -19,6 +19,7 @@
   - pan.spread~ @ch property renamed to @n (number of channels)
   - random.atom @a property renamed to @value
   - list.repeat @times property renamed to @n
+  - fx.zita_rev1~ @delay property renamed to @predelay
 
 ## [0.9.5]
 ### Added: